Adding solution file for the AirQualityMX repo
Solution file for dotnet project is currently listing 3(will grow) main projects
located in the following directories.
- airquality/src: Will hold projects which convert to class libraries
and will have code to be executed in production as part of the "system"
- airquality/tst: holds code that will "test" the "system"
- airqualityCdk: holds code that will represent the cloud infastructure,
such as the "stacks" of the system. Used to automate deployment and
manage resources in the system.

Note: although "Cdk" is not required to build/test the code itself,
having it in the root sln file will let VSCode to properly apply C#
Intellisense